+++ commands/eval.js
@@ -1,0 +1,7 @@
+module.exports = {
+	name: 'eval',
+	description: 'Evaluates the code - DEVELOPER only',
+	execute(message, args) {
+		message.channel.send('Pong.');
+	},
+};
