+++ data.js
@@ -1,0 +1,86 @@
+module.exports = {
+        execute() {
+            const faces = [
+                "(・\`ω´・)",
+                ";;w;;",
+                "owo",
+                "UwU",
+                ">w<",
+                "^w^",
+                "✧w✧",
+                "♥w♥",
+                "(˘³˘)",
+                "(。U⁄ ⁄ω⁄ ⁄ U。)",
+                "(ᵘʷᵘ)",
+                "(ᵕᴗ ᵕ\⁎)",
+                "uwU",
+                "◔w◔",
+                "⓪w⓪",
+                "‿︵𝓇𝒶𝓌𝓇‿︵ ʘwʘ",
+                "øwø",
+                "ÓwÓ",
+                "ÕwÕ",
+                "@w@",
+                "ᅌwᅌ",
+                "ʘwʘ",
+                "(✿◠‿◠)",
+                "(●´ω｀●)",
+                "(づ｡◕‿‿◕｡)づ",
+                "≧◡≦",
+                "(◡‿◡✿)",
+                "(\*^ -^\*)",
+                "(∪ ◡ ∪)",
+                "(✿◠‿◠)",
+                "╰(◡‿◡✿╰)",
+                "(ﾉ◕ヮ◕)ﾉ\*:･ﾟ✧",
+                "(￣ｰ￣)",
+                "ヽ(゜∇゜)ノ",
+                "(◕ω◕✿)",
+                "(〃^∇^)ﾉ",
+                "(´｡• ᵕ •｡\`)",
+                "ヽ(>∀<☆)ノ",
+                "ヽ(\*・ω・)ﾉ",
+                "☆ ～('▽^人)",
+                "(´ ω \`♡)",
+                "(๑˃ᴗ˂)ﻭ",
+                "( ´ ▽ \` ).｡ｏ♡",
+                "╰(\*´︶`\*)╯♡",
+                "ヽ(♡‿♡)ノ",
+                "( ´ ∀ \`)ノ～ ♡",
+                "♡ ～('▽^人)",
+                "( ´ ▽ \` ).｡ｏ♡",
+                "Σ>―(〃°ω°〃)♡→",
+                "(´,,•ω•,,)♡",
+                "( ˘⌣˘)♡(˘⌣˘ )",
+                "(„ಡωಡ„)",
+                "(ノ\*°▽°\*)",
+                "(｡･ω･｡)ﾉ♡",
+                "(=^･ω･^=)",
+                "╰(◡‿◡✿╰)",
+                "(´･ω･\`)",
+                "(=^-ω-^=)",
+                "ヽ(=^･ω･^=)丿",
+                "ʚ(\*´꒳\`\*)ɞ",
+                "(´♡ω♡\`)",
+                "★\~(◡﹏◕✿)",
+                "★\~(◡ω◕✿)",
+                "★\~(◡﹏◡✿)",
+                "★\~(◠︿⊙✿)",
+                "｡◕ ‿ ◕｡",
+                "(◠︿◠✿)"
+            ];
+
+            const statuses = [
+                "with uwu faces",
+                'a game called "OwO what\'s this?!?"',
+                "*notices shitpost* uwu what's this?",
+                "@mention me to uwu-ify messages",
+                "with s-senpai~~",
+                "Coded by code monkeys at uwu headquawews",
+                "Hewwo dewr~!",
+                "Stay safe! uwu",
+                "Stay home!",
+                "Remember to wash your hands!"
+            ];
+
+        }
