+++ package.json
@@ -1,22 +1,25 @@
 {
   "name": "uwutranslator",
-  "version": "1.0.0",
-  "description": "Discord bot that uwu-ifies your messages. uwu",
+  "version": "2.0.0",
+  "description": "Discord bot, which uwu-ifies your messages",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "keywords": [],
-  "author": "TheOnlyGhostwolf",
-  "homepage": "https://uwutranslator.ghostwolf.me",
-  "repository": "https://github.com/TheOnlyGhostwolf/uwutranslator",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/TheOnlyGhostwolf/uwutranslator.git"
+  },
+  "author": "Paulius Gečas",
   "license": "GPL-3.0-only",
+  "bugs": {
+    "url": "https://github.com/TheOnlyGhostwolf/uwutranslator/issues"
+  },
+  "homepage": "https://github.com/TheOnlyGhostwolf/uwutranslator#readme",
   "dependencies": {
-    "discord.js": "^11.5.1",
-    "dotenv": "8.2.0",
-    "express-ping": "1.4.0",
     "axios": "^0.19.2",
-    "dblapi.js": "^2.3.0"
+    "dblapi.js": "^2.3.0",
+    "discord.js": "12.1.1",
+    "dotenv": "8.2.0"
   }
 }
